Use assertEquals for better error messages
when the test fails we only get expected true but was false with
assertTrue. We loose information in these specific cases because knowing
what the actual value was might help us fix the issue faster. Locally
but also more importantly on a PR or push to master when looking at the
CI logs.