TRUNK-5816 Migrate MediaHandlerTest to JUnit5
Since PowerMock is not supported on JUnit 5 we need to convert it to an
integration test. If the handler would have the AdministrationService as
a dependency passed like for example the constructor we would enable
unit testing. Reaching for services via the Context is a bad practice
that we need to get rid of for exactly such cases.