TRUNK-5816 Disable failing HandlerTests on Windows
only for now. We need to replicate what developers are seeing on Windows
see discussion at https://talk.openmrs.org/t/tests-failing-in-core-on-windows/29471/16
the tests still run in our pipelines (Travis/Bamboo) which ensures that
the behaviour stays the same. This will at least unblock developers on
Windows until we find the root cause.