TRUNK-5816 Run StartModuleAnnotationTest
it was not run since all tests in the test folder are excluded from the
surefire plugin in the api

https://github.com/openmrs/openmrs-core/blob/9d4e3d4a48f8505ec009988658678e087a080dff/api/pom.xml#L363-L367
Most likely so that the BaseContext...Test classes dont get picked up.
Not sure if they would actually run tough since they have no @Test
annoation. We can double check that.

Moving the test to the annotation package since there is already a
StartModuleAnnotation test