TRUNK-5816 Do not allow JUnit 4 tests to be added to the api
we allow only a handful that we need to ensure module devs can still
write JUnit 4 tests. If there is a bug and you would need to add another
one please name them with JUnit4 like the examples. This will allow us
to remove them once we are on JUnit 5. Every other new test should be written
in JUnit 5 :)