Autonatically update maven dependencies
We are updating all libraries in core during the GSoC so that core is up to date. We should use the opportunity and automation to keep core up to date going forward. This will prevent us from having to do a big bang update by continously putting in a little effort.